move all vector generation to vectors_test.go
also streamline vectors file structure
